Moved Spec validation into its own separate class, SpecValidator
