+++ src/mlte/spec/__init__.py
@@ -1,5 +1,6 @@
 from .spec import Spec
 from .bound_spec import BoundSpec
 from .condition import Condition
+from .spec_validator import SpecValidator
 
-__all__ = ["Spec", "BoundSpec", "Condition"]
+__all__ = ["Spec", "BoundSpec", "Condition", "SpecValidator"]
